move helper functions to new module
